Move analytical fairness computations into theory submodule
